Reorganized a bunch of the code
- Moved settings into their own file
- Created the BRSW2_CONST class for storing const values
- Moved a bunch of variables into said class
