+++ scripts/brsw2-config.js
@@ -5,130 +5,18 @@ export const MODULE_NAME = "betterrolls-swade2";
 //export const MODULE_SHORT_TITLE = "BR2";
 //export const MODULE_PATH = "modules/betterrolls-swade2";
 
-export const ATTRIBUTES = ["agility", "smarts", "spirit", "strength", "vigor"];
-
-export const ARCANE_SKILLS = [
-  "faith",
-  "focus",
-  "spellcasting",
-  `glaube`,
-  "fokus",
-  "zaubern",
-  "druidism",
-  "elementalism",
-  "glamour",
-  "heahwisardry",
-  "hrimwisardry",
-  "solar magic",
-  "song magic",
-  "soul binding",
-  "artificer",
-  "astrology",
-  "dervish",
-  "divination",
-  "jinn binding",
-  "khem-hekau",
-  "mathemagic",
-  "sand magic",
-  "sha'ir",
-  "ship magic",
-  "ushabti",
-  "wizir magic",
-  "word magic",
-  "druidenmagie",
-  "elementarmagie",
-  "heahmagie",
-  "hrimmagie",
-  "gesangsmagie",
-  "psionics",
-  "psiónica",
-  "psionica",
-  "fe",
-  "hechicería",
-  "hechiceria",
-  "foi",
-  "magie",
-  "science étrange",
-  "science etrange",
-  "élémentalisme",
-  "elementalisme",
-  "druidisme",
-  "magie solaire",
-  "weird science",
-  "voidomancy",
-];
-
-export const FIGHTING_SKILLS = [
-  "fighting",
-  "kämpfen",
-  "pelear",
-  "combat",
-  "lutar",
-  "combattere",
-];
-
-export const SHOOTING_SKILLS = [
-  "shooting",
-  "schiessen",
-  "disparar",
-  "tir",
-  "atirar",
-  "sparare",
-];
-
-export const THROWING_SKILLS = [
-  "athletics",
-  "athletik",
-  "atletismo",
-  "athletisme",
-  "athlétisme",
-  "★ athletics",
-  "atletica",
-];
-
-export const UNTRAINED_SKILLS = [
-  "untrained",
-  "untrainiert",
-  "desentrenada",
-  "non entraine",
-  "non entrainé",
-  "unskilled",
-  "unskilled attempt",
-  "(unskilled)",
-];
-
-export const ARCANE_MASTERY_EDGES = [
-  "BRSW.EdgeName.ArcaneMastery",
-  "BRSW.EdgeName.DivineMastery",
-  "BRSW.EdgeName.EpicMastery",
-  "BRSW.EdgeName.MasterOfMagic",
-  "BRSW.EdgeName.MasterPsionics",
-];
-
 export const MAX_NOPP_PENALTY_ACTION = 6;
 
-/**
-/ Translation map for attributes
-*/
-
-export const ATTRIBUTES_TRANSLATION_KEYS = {
-  agility: "SWADE.AttrAgi",
-  smarts: "SWADE.AttrSma",
-  spirit: "SWADE.AttrSpr",
-  strength: "SWADE.AttrStr",
-  vigor: "SWADE.AttrVig",
-};
-
 export const SETTING_KEYS = {
-  auto_check_extra_fumbles: "auto_check_extra_fumbles",
-  enabledOptionalRules: "optional_rules_enabled",
-  telemetryOptOut: "telemetryOptOut",
-  telemetryWorldInstallId: "telemetryWorldInstallId",
-  worldSettings: "world_settings",
+    autoCheckExtraCritFailures: "auto_check_extra_fumbles",
+    enabledOptionalRules: "optional_rules_enabled",
+    telemetryOptOut: "telemetryOptOut",
+    telemetryWorldInstallId: "telemetryWorldInstallId",
+    worldSettings: "world_settings",
 };
 
 export const USER_FLAGS = {
-  userSettings: "user_settings",
+    userSettings: "user_settings",
 };
 
 export const WORLD_SETTINGS = {};
