+++ scripts/utils.js
@@ -1,4 +1,5 @@
 import * as BRSW2_CONFIG from "./brsw2-config.js";
+import { BRSW2_CONST } from "./brsw2-const.js";
 
 // Utility functions that can be used out of the module
 /* globals ChatMessage, game, console, foundry, ClientSetting, CONFIG */
@@ -403,9 +404,9 @@ export class Utils {
 
         if (!trait) {
             // Time to check for an attribute
-            for (const attribute of BRSW2_CONFIG.ATTRIBUTES) {
+            for (const attribute of BRSW2_CONST.ATTRIBUTES) {
                 const translation = game.i18n.localize(
-                    BRSW2_CONFIG.ATTRIBUTES_TRANSLATION_KEYS[attribute],
+                    BRSW2_CONST.ATTRIBUTES_TRANSLATION_KEYS[attribute],
                 );
                 if (traitLower === translation.toLowerCase()) {
                     trait = { system: structuredClone(actor.system.attributes[attribute]) };
@@ -417,7 +418,7 @@ export class Utils {
         if (!trait) {
             // No skill was found, we try to find untrained
             trait = Utils.findFirstSkillInActor(actor, [
-                ...BRSW2_CONFIG.UNTRAINED_SKILLS,
+                ...BRSW2_CONST.UNTRAINED_SKILLS,
                 game.i18n.localize("BRSW.SkillName.UnskilledAttempt").toLowerCase(),
             ]);
         }
@@ -476,24 +477,24 @@ export class Utils {
         // If there is no skill anyway, we are left to guessing
         let skill;
         if (item.type === "power") {
-            skill = Utils.findFirstSkillInActor(actor, BRSW2_CONFIG.ARCANE_SKILLS);
+            skill = Utils.findFirstSkillInActor(actor, BRSW2_CONST.ARCANE_SKILLS);
         } else if (item.type === "weapon") {
             if (parseInt(item.system.range) > 0) {
                 // noinspection JSUnresolvedVariable
                 if (item.system.damage.includes("str")) {
                     skill = Utils.findFirstSkillInActor(actor, [
-                        ...BRSW2_CONFIG.THROWING_SKILLS,
+                        ...BRSW2_CONST.THROWING_SKILLS,
                         game.i18n.localize("BRSW.SkillName.Athletics").toLowerCase(), // add localization
                     ]);
                 } else {
                     skill = Utils.findFirstSkillInActor(actor, [
-                        ...BRSW2_CONFIG.SHOOTING_SKILLS,
+                        ...BRSW2_CONST.SHOOTING_SKILLS,
                         game.i18n.localize("BRSW.SkillName.Shooting").toLowerCase(), // add localization
                     ]);
                 }
             } else {
                 skill = Utils.findFirstSkillInActor(actor, [
-                    ...BRSW2_CONFIG.FIGHTING_SKILLS,
+                    ...BRSW2_CONST.FIGHTING_SKILLS,
                     game.i18n.localize("BRSW.SkillName.Fighting").toLowerCase(), // bag add localization
                 ]);
             }
@@ -501,7 +502,7 @@ export class Utils {
 
         if (skill === undefined) {
             skill = Utils.findFirstSkillInActor(actor, [
-                ...BRSW2_CONFIG.UNTRAINED_SKILLS,
+                ...BRSW2_CONST.UNTRAINED_SKILLS,
                 game.i18n.localize("BRSW.SkillName.UnskilledAttempt").toLowerCase(),
             ]);
         }
@@ -538,7 +539,7 @@ export class Utils {
         }
 
         const configured_skill_name = game.settings.get("swade", "parryBaseSkill").toLowerCase();
-        const fightingNames = BRSW2_CONFIG.FIGHTING_SKILLS;
+        const fightingNames = BRSW2_CONST.FIGHTING_SKILLS;
         fightingNames.push(configured_skill_name);
         return fightingNames.includes(skill.name.toLowerCase());
     }
@@ -552,7 +553,7 @@ export class Utils {
         if (!skill) return false;
         if (skill.system.swid === "shooting") return true;
 
-        const shootingNames = BRSW2_CONFIG.SHOOTING_SKILLS;
+        const shootingNames = BRSW2_CONST.SHOOTING_SKILLS;
         shootingNames.push(game.i18n.localize("BRSW.SkillName.Shooting"));
         return shootingNames.includes(skill.name.toLowerCase());
     }
@@ -566,7 +567,7 @@ export class Utils {
         if (!skill) return false;
         if (skill.system.swid === "athletics") return true;
 
-        const throwingNames = BRSW2_CONFIG.THROWING_SKILLS;
+        const throwingNames = BRSW2_CONST.THROWING_SKILLS;
         throwingNames.push(game.i18n.localize("BRSW.SkillName.Athletics"));
         return throwingNames.includes(skill.name.toLowerCase());
     }
@@ -606,7 +607,7 @@ export class Utils {
     }
 
     static actorHasArcaneMastery(actor) {
-        const edgeNames = BRSW2_CONFIG.ARCANE_MASTERY_EDGES.map((edge) => game.i18n.localize(edge).toLowerCase());
+        const edgeNames = BRSW2_CONST.ARCANE_MASTERY_EDGES.map((edge) => game.i18n.localize(edge).toLowerCase());
         const edge = actor?.items.find((item) => {
             return (
                 item.type === "edge" &&
