+++ templates/settings_config.hbs
@@ -1,6 +1,6 @@
 <form>
     <nav class="sheet-tabs tabs" data-group="main" aria-role="Form Tab Navigation">
-        {{#if can_modify_world}}
+        {{#if canModifyWorld}}
             <a class="item active" data-tab="world"><i class="fas fa-globe"></i> {{localize "BRSW.Settings.Tabs.World"}}</a>
         {{/if}}
         <a class="item" data-tab="user"><i class="fas fa-user"></i> {{localize "BRSW.Settings.Tabs.User"}}</a>
