+++ templates/settings_config/world_tab.hbs
@@ -1,5 +1,5 @@
 <section class="tab active scrollable" data-group="primary" data-tab="world">
-    {{#if can_modify_world}}
+    {{#if canModifyWorld}}
         {{#each worldSettings}}
             {{>"modules/betterrolls-swade2/templates/setting_partial.hbs"}}
         {{/each}}
